+++ src/app/dashboard/components/UserDash.js
@@ -10,8 +10,8 @@ import {
 
 export default function UserDash() {
   return (
-    <section className="userDash--container flex bg-brand-primary/60 rounded-lg mb-6 mx-auto md:max-w-3xl xl:max-w-6xl">
-      <div className="bg-gray-200/5 rounded-lg">
+    <section className="userDash--container flex bg-brand-primary/60 rounded-lg mb-6 mx-auto">
+      <div className="bg-gray-200/5 rounded-lg w-full">
         <div className="userDash--infoContainer p-5 flex flex-col items-center md:flex-row md:px-8 md:py-12 md:gap-4">
           <div className="flex mr-auto gap-4">
             <div className="userDash--imageContainer flex items-center">
@@ -28,7 +28,7 @@ export default function UserDash() {
         </div>
         <div className="userDash--linksContainer pb-10 flex flex-col md:px-8 md:gap-2 text-lg font-primary">
           <Link href="/dashboard" className="flex items-center justify-between">
-          <div className="delay-75 duration-300 py-1 rounded-lg w-fill flex items-center gap-3">
+            <div className="delay-75 duration-300 py-1 rounded-lg w-fill flex items-center gap-3">
               <FontAwesomeIcon
                 icon={faUser}
                 className="bg-brand-secondary/5 hover:bg-brand-secondary/10 delay-75 duration-300 text-brand-secondary p-2 rounded-lg w-5 h-5"
@@ -38,7 +38,7 @@ export default function UserDash() {
             <FontAwesomeIcon icon={faAngleRight} className="w-4 h-4" />
           </Link>
           <Link href="/dashboard" className="flex items-center justify-between">
-          <div className="delay-75 duration-300 py-1 rounded-lg w-fill flex items-center gap-3">
+            <div className="delay-75 duration-300 py-1 rounded-lg w-fill flex items-center gap-3">
               <FontAwesomeIcon
                 icon={faCalendarCheck}
                 className="bg-brand-secondary/5 hover:bg-brand-secondary/10 delay-75 duration-300 text-brand-secondary p-2 rounded-lg w-5 h-5"
